DS-1367 New Missing Service Type (#861)
# Task Branch Pull Request

**<https://nhsd-jira.digital.nhs.uk/browse/DS-1367>**

## Description of Changes

This PR refactors the service matcher as well as adding a new log to the
missing service type report if the main pharmacy doesn't exist and
palliative care should be switched on.
